requirements.txt
pyyaml>=6.0.1
colorama
pyperclip
typing-extensions>=4.7.1 